+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added launching simulation.
 - Added generating random seed using OS-specific randomness source.
 - Added simulation launcher console script.
+- Added random seed generator console script.
 - Added hierarchy of SimTools exceptions.
 - Added package installer.
 - Added package description.